Revert "WIP add cloudflared local, cloudflared docker and jurassic tube"
This reverts commit 2221e505fd7aab93590ee7f5b6642dfab559a05d.
