+++ src/src/Environment/Environments/E2E/E2EEnvironment.php
@@ -55,7 +55,7 @@ class E2EEnvironment extends Environment {
 			$this->env_info->nginx_port = '80';
 			$nginx_container            = "qit_env_nginx_{$this->env_info->env_id}";
 
-			$site_url = App::make( TunnelRunner::class )->get_tunnel()->start_tunnel( "http://$nginx_container", $this->env_info->env_id, $this->env_info->docker_network );
+			$site_url = App::make( TunnelRunner::class )->start_tunnel( "http://$nginx_container", $this->env_info->env_id, $this->env_info->docker_network );
 
 			$this->env_info->domain     = parse_url( $site_url, PHP_URL_HOST );
 			$this->env_info->nginx_port = (string) parse_url( $site_url, PHP_URL_PORT );
