+++ src/src/Tunnel/TunnelRunner.php
@@ -2,49 +2,81 @@
 
 namespace QIT_CLI\Tunnel;
 
-use QIT_CLI\App;
-use QIT_CLI\Cache;
+use QIT_CLI\Environment\Docker;
+use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Process\Process;
 
 class TunnelRunner {
-	/** @var Cache */
-	protected $cache;
+	/** @var Docker */
+	protected $docker;
 
-	public function __construct( Cache $cache ) {
-		$this->cache = $cache;
+	/** @var OutputInterface */
+	protected $output;
+
+	public function __construct( Docker $docker, OutputInterface $output ) {
+		$this->docker = $docker;
+		$this->output = $output;
 	}
 
 	public function start_tunnel( string $local_url, string $env_id, string $network_name ): string {
-		return $this->get_tunnel()->start_tunnel( $local_url, $env_id, $network_name );
-	}
+		if ( ! empty( getenv( 'QIT_TUNNEL_REGION' ) ) ) {
+			// Validate $region is a-z.
+			if ( ! preg_match( '/^[a-z]+$/', getenv( 'QIT_TUNNEL_REGION' ) ) ) {
+				throw new \InvalidArgumentException( 'Invalid region specified.' );
+			}
+			$region = '--region=' . getenv( 'QIT_TUNNEL_REGION' );
+		} else {
+			// If not defined, Cloudflare will allocate the fastest region to the requester.
+			// Some payment gateways might require "us" region, so we allow them to override it if needed.
+			$region = '';
+		}
 
-	public function stop_tunnel( string $env_id ): void {
-		$this->get_tunnel()->stop_tunnel( $env_id );
-	}
+		$local_url             = escapeshellarg( $local_url );
+		$docker_container_name = "qit_env_tunnel_$env_id";
 
-	public function set_tunnel( string $tunnel_id ): void {
-		$this->cache->set( 'tunnel', $tunnel_id, - 1 );
-	}
+		// Start the Docker container in detached mode.
+		exec( "{$this->docker->find_docker()} run -d --network=$network_name --name=$docker_container_name cloudflare/cloudflared:latest tunnel --no-autoupdate $region --url=$local_url", $run_output, $run_return );
 
-	public function get_tunnel(): Tunnel {
-		$tunnel_id = $this->cache->get( 'tunnel' );
+		if ( $run_return !== 0 ) {
+			$error_output = implode( "\n", $run_output );
+			throw new \RuntimeException( "Failed to start tunnel container: $error_output" );
+		}
 
-		// If no tunnel is previously configured and the Cloudflared Docker tunnel is available, use it by default.
-		// This will be mostly used in CI.
-		if ( empty( $tunnel_id ) ) {
-			if ( App::make( CloudflaredDockerTunnel::class )->is_available() ) {
-				return App::make( CloudflaredDockerTunnel::class );
+		// Initialize variables for capturing the domain.
+		$domain      = null;
+		$start_time  = time();
+		$timeout     = 60; // seconds.
+		$logs_output = [];
+
+		// Loop to check the container logs for the domain.
+		while ( ( time() - $start_time ) < $timeout ) {
+			exec( "{$this->docker->find_docker()} logs $docker_container_name 2>&1", $logs_output, $logs_return );
+
+			$logs = implode( "\n", $logs_output );
+			if ( preg_match( '#https://[a-zA-Z0-9\-]+\.trycloudflare\.com#', $logs, $matches ) ) {
+				$domain = $matches[0];
+				break;
+			}
+
+			usleep( 500000 ); // 0.5 seconds
+		}
+
+		if ( $this->output->isVeryVerbose() ) {
+			foreach ( $logs_output as $log ) {
+				$this->output->writeln( $log );
 			}
 		}
 
-		switch ( $tunnel_id ) {
-			case 'cloudflared_docker':
-				return App::make( CloudflaredDockerTunnel::class );
-			case 'cloudflared_local':
-				return App::make( CloudflaredLocalTunnel::class );
-			case 'jurassic_tube_local':
-				return App::make( JurassicTubeTunnel::class );
-			default:
-				throw new \InvalidArgumentException( 'Invalid tunnel ID.' );
+		if ( $domain === null ) {
+			static::stop_tunnel( $env_id );
+			throw new \RuntimeException( 'Timed out waiting for tunnel domain.' );
 		}
+
+		return $domain;
+	}
+
+	public static function stop_tunnel( string $env_id ): void {
+		$p = new Process( [ 'docker', 'rm', '-f', "qit_env_tunnel_$env_id" ] );
+		$p->run();
 	}
 }
